fix: moved common code (that required no architecture-specific modifications) outside the IntrinsicDefinition trait
