+++ crates/intrinsic-test/src/common/write_file.rs
@@ -1,6 +1,7 @@
+use super::gen_c::gen_c_program;
+use super::gen_rust::{create_rust_filenames, gen_rust_program};
 use super::intrinsic_helpers::IntrinsicTypeDefinition;
 use crate::common::gen_c::create_c_filenames;
-use crate::common::gen_rust::create_rust_filenames;
 use crate::common::intrinsic::IntrinsicDefinition;
 use crate::common::write_file;
 
@@ -18,9 +19,15 @@ pub fn write_c_testfiles<T: IntrinsicTypeDefinition + Sized>(
         .collect::<Vec<_>>();
     let filename_mapping = create_c_filenames(&intrinsics_name_list);
 
-    intrinsics.iter().for_each(|i| {
-        let c_code =
-            i.generate_c_program(headers, target, c_target, notice, arch_specific_definitions);
+    intrinsics.iter().for_each(|&i| {
+        let c_code = gen_c_program(
+            i,
+            headers,
+            target,
+            c_target,
+            notice,
+            arch_specific_definitions,
+        );
         match filename_mapping.get(&i.name()) {
             Some(filename) => write_file(filename, c_code),
             None => {}
@@ -43,8 +50,8 @@ pub fn write_rust_testfiles<T: IntrinsicTypeDefinition>(
         .collect::<Vec<_>>();
     let filename_mapping = create_rust_filenames(&intrinsics_name_list);
 
-    intrinsics.iter().for_each(|i| {
-        let rust_code = i.generate_rust_program(rust_target, notice, definitions, cfg);
+    intrinsics.iter().for_each(|&i| {
+        let rust_code = gen_rust_program(i, rust_target, notice, definitions, cfg);
         match filename_mapping.get(&i.name()) {
             Some(filename) => write_file(filename, rust_code),
             None => {}
